+++ aws-opentelemetry-distro/tests/amazon/opentelemetry/distro/test_utils.py
@@ -9,10 +9,12 @@ from unittest.mock import MagicMock, patch
 from amazon.opentelemetry.distro._utils import (
     AGENT_OBSERVABILITY_ENABLED,
     AWS_GENAI_CONTENT_EXTRACTION_OPT_OUT,
+    AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN,
     get_aws_region,
     get_aws_session,
     is_agent_observability_enabled,
     is_genai_content_extraction_opted_out,
+    is_genai_latest_llo_content_extraction_opted_in,
     is_installed,
 )
 
@@ -201,3 +203,26 @@ class TestUtils(TestCase):
         if AWS_GENAI_CONTENT_EXTRACTION_OPT_OUT in os.environ:
             del os.environ[AWS_GENAI_CONTENT_EXTRACTION_OPT_OUT]
         self.assertFalse(is_genai_content_extraction_opted_out())
+
+    def test_is_genai_latest_llo_content_extraction_opted_in_various_values(self):
+        os.environ[AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN] = "true"
+        self.assertTrue(is_genai_latest_llo_content_extraction_opted_in())
+
+        os.environ[AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN] = "True"
+        self.assertTrue(is_genai_latest_llo_content_extraction_opted_in())
+
+        os.environ[AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN] = "TRUE"
+        self.assertTrue(is_genai_latest_llo_content_extraction_opted_in())
+
+        os.environ[AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN] = "false"
+        self.assertFalse(is_genai_latest_llo_content_extraction_opted_in())
+
+        os.environ[AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN] = "False"
+        self.assertFalse(is_genai_latest_llo_content_extraction_opted_in())
+
+        os.environ[AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN] = ""
+        self.assertFalse(is_genai_latest_llo_content_extraction_opted_in())
+
+        if AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN in os.environ:
+            del os.environ[AWS_GENAI_LATEST_LLO_CONTENT_EXTRACTION_OPT_IN]
+        self.assertFalse(is_genai_latest_llo_content_extraction_opted_in())